Add pen option button events and minor tweaks
Added the ability to enable the different pen options when clicking the
corresponding button, and added the function to make each cell a random
color with the Random option. Gradient function still a work in progress.

Updated the grid-instruct element to recommend a size less than 50 for
mobile as a test of 100 caused the mobile browser to crash.