Fix gradient color issue, minor tweaks
Fixed an issue with the gradient option where trying to fill in a cell that
had a previous alpha value with a new gradient color would keep the old alpha
value instead of starting back at 0.1.

Changed the default fonts and added fallback, and made the gradient pen icon
easier to see.